requirements.txt
from_root==1.0.2
tqdm==4.64.1
split-folders==0.5.1
boto3==1.24.73
pandas==1.4.4
scikit-learn==1.1.2
mysql-connector-python==8.0.30
pymongo==4.2.0
annoy==1.17.1
pymongo[srv]
torch
torchvision
pillow